Some work done, barely any progress.
As usual. Quitted early because I can, and because I had nothing better to do.

EoN remains a waste of time currently, and I should really just quit making it. Is it really real coding if it's Lua? All I'm doing is putting commands down, dealing with the nonsense, going on with the game, putting commands down, repeat ad nauseum.